2win-mvp/backend/requirements.txt
fastapi>=0.95.2
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
supabase==1.0.3
python-dotenv>=1.0.0
pydantic[email]>=1.10.7
pydantic-settings>=2.0.0
httpx>=0.23.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
numpy>=1.24.0
scikit-learn>=1.3.0
apscheduler>=3.10.0
